Refactor `useDrop` hook to contain droppable functionality.
